+++ src/main/java/net/geforcemods/securitycraft/api/IModuleInventory.java
@@ -154,7 +154,7 @@ public interface IModuleInventory extends IItemHandlerModifiable {
 				continue;
 
 			if (linkable != null)
-				linkable.createLinkedBlockAction(new ILinkedAction.ModuleRemoved(((ModuleItem) module.getItem()).getModuleType(), false), linkable);
+				linkable.propagate(new ILinkedAction.ModuleRemoved(((ModuleItem) module.getItem()).getModuleType(), false), linkable);
 
 			Block.spawnAsEntity(level, pos, module);
 		}
@@ -195,7 +195,7 @@ public interface IModuleInventory extends IItemHandlerModifiable {
 					if (te instanceof LinkableBlockEntity) {
 						LinkableBlockEntity linkable = (LinkableBlockEntity) te;
 
-						linkable.createLinkedBlockAction(new ILinkedAction.ModuleRemoved(((ModuleItem) stack.getItem()).getModuleType(), false), linkable);
+						linkable.propagate(new ILinkedAction.ModuleRemoved(((ModuleItem) stack.getItem()).getModuleType(), false), linkable);
 					}
 				}
 
@@ -232,7 +232,7 @@ public interface IModuleInventory extends IItemHandlerModifiable {
 					if (te instanceof LinkableBlockEntity) {
 						LinkableBlockEntity linkable = (LinkableBlockEntity) te;
 
-						linkable.createLinkedBlockAction(new ILinkedAction.ModuleInserted(copy, (ModuleItem) copy.getItem(), false), linkable);
+						linkable.propagate(new ILinkedAction.ModuleInserted(copy, (ModuleItem) copy.getItem(), false), linkable);
 					}
 				}
 			}
@@ -262,7 +262,7 @@ public interface IModuleInventory extends IItemHandlerModifiable {
 			if (te instanceof LinkableBlockEntity) {
 				LinkableBlockEntity linkable = (LinkableBlockEntity) te;
 
-				linkable.createLinkedBlockAction(new ILinkedAction.ModuleRemoved(((ModuleItem) previous.getItem()).getModuleType(), false), linkable);
+				linkable.propagate(new ILinkedAction.ModuleRemoved(((ModuleItem) previous.getItem()).getModuleType(), false), linkable);
 			}
 		}
 
@@ -274,7 +274,7 @@ public interface IModuleInventory extends IItemHandlerModifiable {
 			if (te instanceof LinkableBlockEntity) {
 				LinkableBlockEntity linkable = (LinkableBlockEntity) te;
 
-				linkable.createLinkedBlockAction(new ILinkedAction.ModuleInserted(stack, (ModuleItem) stack.getItem(), false), linkable);
+				linkable.propagate(new ILinkedAction.ModuleInserted(stack, (ModuleItem) stack.getItem(), false), linkable);
 			}
 		}
 	}