+++ src/main/java/net/geforcemods/securitycraft/blockentities/LaserBlockBlockEntity.java
@@ -165,7 +165,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements IInven
 		}
 
 		excludedTEs.add(this);
-		createLinkedBlockAction(action, excludedTEs);
+		propagate(action, excludedTEs);
 	}
 
 	@Override
@@ -452,7 +452,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements IInven
 			for (ModuleType type : thatInsertedModules) {
 				bothInsertedModules.put(that.getModule(type).copy(), that.isModuleEnabled(type));
 				that.removeModule(type, false);
-				createLinkedBlockAction(new ILinkedAction.ModuleRemoved(type, false), that);
+				propagate(new ILinkedAction.ModuleRemoved(type, false), that);
 			}
 
 			readOptions(that.writeOptions(new NBTTagCompound()));
@@ -464,9 +464,9 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements IInven
 				ModuleType type = item.getModuleType();
 
 				insertModule(entry.getKey(), false);
-				createLinkedBlockAction(new ILinkedAction.ModuleInserted(module, item, false), this);
+				propagate(new ILinkedAction.ModuleInserted(module, item, false), this);
 				toggleModuleState(type, entry.getValue());
-				createLinkedBlockAction(new ILinkedAction.ModuleInserted(module, item, true), this);
+				propagate(new ILinkedAction.ModuleInserted(module, item, true), this);
 			}
 		}